+++ sources/framework/sandbox/sandbox.js
@@ -29,6 +29,7 @@
             var me = this;
 
             me._core = core;
+            this._map = Oskari.clazz.create('Oskari.mapframework.domain.Map');
 
             /*
              * All registered listeners in map key: event name value:
@@ -513,18 +514,6 @@
                 }
                 return this._modulesByName[name];
             },
-
-            /**
-             * @method getRequestParameter
-             * Returns a request parameter from query string
-             * http://javablog.info/2008/04/17/url-request-parameters-using-javascript/
-             * @param {String} name - parameter name
-             * @return {String} value for the parameter or null if not found
-             */
-            getRequestParameter: function (name) {
-                return this._core.getRequestParameter(name);
-            },
-
 
             /**
              * @method getBrowserWindowSize